S27 Modularizacion con export
