+++ readme.md
@@ -1,3 +1,15 @@
 # PWPC I
 Este sera el proyecto de la materia
 de progrmacaion web en pila completa I
+
+Browsers => Ram [GLOBAL          ]
+
+ES3 === Es4X
+2009 => CommonJs => NODE
+require() Permite importar modulos
+exports Permite exportar modulos
+
+ESS
+"export"
+1. Exportaciones nombradas
+2. Exportaciones por defecto
